Use utils file for tests
